refactor(SPV-1173): simplify examples dir hierarchy.
