--- docs/examples/api/v1/users/contacts/contact_remove/contact_remove.go
@@ -1,26 +1,0 @@
-package main
-
-import (
-	"context"
-	"fmt"
-	"log"
-
-	wallet "github.com/bitcoin-sv/spv-wallet-go-client"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples/exampleutil"
-)
-
-func main() {
-	spv, err := wallet.NewWithXPriv(exampleutil.ExampleConfig, examples.XPriv)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	paymail := "john.doe@example.com"
-	err = spv.RemoveContact(context.Background(), paymail)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Println(fmt.Sprintf("\n[HTTP DELETE] Remove contact - api/v1/contacts/%s", paymail))
-}
