--- docs/examples/api/v1/users/contacts/contact_upsert/contact_upsert.go
@@ -1,33 +1,0 @@
-package main
-
-import (
-	"context"
-	"fmt"
-	"log"
-
-	wallet "github.com/bitcoin-sv/spv-wallet-go-client"
-	"github.com/bitcoin-sv/spv-wallet-go-client/commands"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples/exampleutil"
-)
-
-func main() {
-	spv, err := wallet.NewWithXPriv(exampleutil.ExampleConfig, examples.XPriv)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	paymail := "john.doe@example.com"
-	contact, err := spv.UpsertContact(context.Background(), commands.UpsertContact{
-		FullName: "John Doe",
-		Metadata: map[string]any{
-			"key": "value",
-		},
-		Paymail: paymail,
-	})
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	exampleutil.Print(fmt.Sprintf("[HTTP PUT] Upsert contact - api/v1/contacts/%s", paymail), contact)
-}
