--- docs/examples/api/v1/users/contacts/fetch_contact/fetch_contact.go
@@ -1,26 +1,0 @@
-package main
-
-import (
-	"context"
-	"fmt"
-	"log"
-
-	wallet "github.com/bitcoin-sv/spv-wallet-go-client"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples/exampleutil"
-)
-
-func main() {
-	spv, err := wallet.NewWithXPriv(exampleutil.ExampleConfig, examples.XPriv)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	paymail := "john.doe@example.com"
-	contact, err := spv.ContactWithPaymail(context.Background(), paymail)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	exampleutil.Print(fmt.Sprintf("[HTTP GET] Contact with paymail - api/v1/contacts/%s", paymail), contact)
-}
