--- docs/examples/api/v1/users/contacts/fetch_contacts/fetch_contacts.go
@@ -1,24 +1,0 @@
-package main
-
-import (
-	"context"
-	"log"
-
-	wallet "github.com/bitcoin-sv/spv-wallet-go-client"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples/exampleutil"
-)
-
-func main() {
-	spv, err := wallet.NewWithXPriv(exampleutil.ExampleConfig, examples.XPriv)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	page, err := spv.Contacts(context.Background())
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	exampleutil.Print("[HTTP GET] Contacts page - api/v1/contacts", page)
-}
