--- docs/examples/api/v1/users/contacts/unconfirm_contact/unconfirm_contact.go
@@ -1,23 +1,0 @@
-package main
-
-import (
-	"context"
-	"log"
-
-	wallet "github.com/bitcoin-sv/spv-wallet-go-client"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples/exampleutil"
-)
-
-func main() {
-	spv, err := wallet.NewWithXPriv(exampleutil.ExampleConfig, examples.XPriv)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	paymail := "john.doe@example.com"
-	err = spv.UnconfirmContact(context.Background(), paymail)
-	if err != nil {
-		log.Fatal(err)
-	}
-}
