--- docs/examples/api/v1/users/invitations/reject_invitation/reject_invitation.go
@@ -1,26 +1,0 @@
-package main
-
-import (
-	"context"
-	"fmt"
-	"log"
-
-	wallet "github.com/bitcoin-sv/spv-wallet-go-client"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples/exampleutil"
-)
-
-func main() {
-	spv, err := wallet.NewWithXPriv(exampleutil.ExampleConfig, examples.XPriv)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	paymail := "john.doe@example.com"
-	err = spv.RejectInvitation(context.Background(), paymail)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	fmt.Println(fmt.Sprintf("\n[HTTP DELETE] Reject contact invitation - api/v1/invitations/%s", paymail))
-}
