--- docs/examples/api/v1/users/transactions/fetch_transactions/fetch_transactions.go
@@ -1,24 +1,0 @@
-package main
-
-import (
-	"context"
-	"log"
-
-	wallet "github.com/bitcoin-sv/spv-wallet-go-client"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples/exampleutil"
-)
-
-func main() {
-	spv, err := wallet.NewWithXPriv(exampleutil.ExampleConfig, examples.XPriv)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	page, err := spv.Transactions(context.Background())
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	exampleutil.Print("[HTTP GET] Transaction page - api/v1/transactions", page)
-}
