--- docs/examples/api/v1/users/xpub/fetch_xpub/fetch_xpub.go
@@ -1,24 +1,0 @@
-package main
-
-import (
-	"context"
-	"log"
-
-	wallet "github.com/bitcoin-sv/spv-wallet-go-client"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples"
-	"github.com/bitcoin-sv/spv-wallet-go-client/docs/examples/exampleutil"
-)
-
-func main() {
-	spv, err := wallet.NewWithXPriv(exampleutil.ExampleConfig, examples.XPriv)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	xPub, err := spv.XPub(context.Background())
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	exampleutil.Print("[HTTP GET] Current user xPub - api/v1/users/current", xPub)
-}
