+++ examples/fetch_contact_by_paymail/fetch_contact_by_paymail.go
@@ -1,0 +1,26 @@
+package main
+
+import (
+	"context"
+	"fmt"
+	"log"
+
+	wallet "github.com/bitcoin-sv/spv-wallet-go-client"
+	"github.com/bitcoin-sv/spv-wallet-go-client/examples"
+	"github.com/bitcoin-sv/spv-wallet-go-client/examples/exampleutil"
+)
+
+func main() {
+	spv, err := wallet.NewWithXPriv(exampleutil.ExampleConfig, examples.XPriv)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	paymail := "john.doe@example.com"
+	contact, err := spv.ContactWithPaymail(context.Background(), paymail)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	exampleutil.Print(fmt.Sprintf("[HTTP GET] Contact by paymail - api/v1/contacts/%s", paymail), contact)
+}
