+++ examples/fetch_shared_config/fetch_shared_config.go
@@ -1,0 +1,24 @@
+package main
+
+import (
+	"context"
+	"log"
+
+	wallet "github.com/bitcoin-sv/spv-wallet-go-client"
+	"github.com/bitcoin-sv/spv-wallet-go-client/examples"
+	"github.com/bitcoin-sv/spv-wallet-go-client/examples/exampleutil"
+)
+
+func main() {
+	spv, err := wallet.NewWithXPriv(exampleutil.ExampleConfig, examples.XPriv)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	res, err := spv.SharedConfig(context.Background())
+	if err != nil {
+		log.Fatal()
+	}
+
+	exampleutil.Print("[HTTP GET] Shared config - api/v1/configs/shared", res)
+}
